Rename common test source file
Help make the subsequent commit diff less noisy.
